+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/track/ContradictionTrack.groovy
@@ -1,0 +1,8 @@
+package ru.vyarus.java.generics.resolver.inlying.support.track
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 10.05.2018
+ */
+class ContradictionTrack<P> extends Target<Integer> {
+}