+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/track/MultiTier.groovy
@@ -1,0 +1,10 @@
+package ru.vyarus.java.generics.resolver.inlying.support.track
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 09.05.2018
+ */
+class MultiTier<A,B> extends Tier2<A> {
+}
+
+class Tier2<P> extends Target<P> {}