+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/track/MultipleGenerics.groovy
@@ -1,0 +1,8 @@
+package ru.vyarus.java.generics.resolver.inlying.support.track
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 09.05.2018
+ */
+class MultipleGenerics<A, B> extends Target<A> {
+}