+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/track/NestedRight.groovy
@@ -1,0 +1,11 @@
+package ru.vyarus.java.generics.resolver.inlying.support.track
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 10.05.2018
+ */
+class NestedRight<K> extends Target<ArrayList<K>> {
+
+    // just declare type here, for simplicity
+    Target<List<String>> target
+}