+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/track/Source.groovy
@@ -1,0 +1,11 @@
+package ru.vyarus.java.generics.resolver.inlying.support.track
+
+import ru.vyarus.java.generics.resolver.inlying.support.track.Target
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 09.05.2018
+ */
+class Source {
+    Target<String> target
+}